requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
addict=2.4.0=pypi_0
aiosignal=1.2.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
attrs=21.4.0=pypi_0
axial-positional-embedding=0.2.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=22.10.0=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
bottleneck=1.3.4=py37h67323c0_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py37h9ed2024_1003
ca-certificates=2022.12.7=h033912b_0
captum=0.6.0=pypi_0
certifi=2022.9.24=pypi_0
cffi=1.15.0=py37hc55c11b_1
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cmake=3.25.0=pypi_0
colorama=0.4.5=pypi_0
comm=0.1.4=pypi_0
commonmark=0.9.1=pypi_0
configargparse=1.5.3=pypi_0
cryptography=37.0.1=py37hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
dash=2.7.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.5.1=py37he9d5cce_0
decorator=5.1.1=pyhd8ed1ab_0
deepsnap=0.2.1=pypi_0
deprecated=1.2.13=pypi_0
dill=0.3.7=pypi_0
dimod=0.12.3=pypi_0
diskcache=5.4.0=pypi_0
distlib=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
docutils=0.19=pypi_0
dwave-cloud-client=0.10.4=pypi_0
dwave-greedy=0.3.0=pypi_0
dwave-hybrid=0.6.10=pypi_0
dwave-inspector=0.4.1=pypi_0
dwave-inspectorapp=0.3.1=pypi_0
dwave-neal=0.6.0=pypi_0
dwave-networkx=0.8.13=pypi_0
dwave-ocean-sdk=6.3.0=pypi_0
dwave-preprocessing=0.5.4=pypi_0
dwave-samplers=1.0.0=pypi_0
dwave-system=1.18.0=pypi_0
dwave-tabu=0.5.0=pypi_0
dwavebinarycsp=0.2.0=pypi_0
einops=0.5.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
fasteners=0.18=pypi_0
fastjsonschema=2.16.2=pypi_0
filelock=3.7.1=pypi_0
flask=2.2.2=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.3.0=pypi_0
future=0.18.2=py37_1
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
graphgym=0.4.0=pypi_0
grpcio=1.43.0=pypi_0
gurobipy=10.0.3=pypi_0
h5py=3.8.0=pypi_0
homebase=1.0.1=pypi_0
hydra-core=1.2.0=pypi_0
idna=3.4=pypi_0
imageio=2.25.0=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.7.1=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.15.0=pyh736e0ef_0
ipython=7.33.0=py37hf985489_0
ipywidgets=8.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jaraco-classes=3.2.3=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
jsonschema=4.6.0=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.11.1=py37hecd8cb5_0
jupyterlab-widgets=3.0.9=pypi_0
keopscore=2.1=pypi_0
keyring=23.9.3=pypi_0
kiwisolver=1.4.2=py37he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=hbcb3906_1
libsvm=3.23.0.4=pypi_0
libtiff=4.2.0=hdb42f99_1
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
littleutils=0.2.2=py_0
llvm-openmp=12.0.0=h0dcd299_1
llvmlite=0.38.1=pypi_0
local-attention=1.4.4=pypi_0
loguru=0.6.0=pypi_0
lsystemrenderer=0.0.9=pypi_0
lz4-c=1.9.3=h23ab428_1
markupsafe=2.1.1=py37hca72f7f_0
mathutils=2.78=pypi_0
matplotlib=3.5.1=py37hecd8cb5_1
matplotlib-base=3.5.1=py37hfb0c5b7_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
minorminer=0.2.11=pypi_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py37h9ed2024_0
mkl_fft=1.3.1=py37h4ab4a9b_0
mkl_random=1.2.2=py37hb2f4e1b_0
more-itertools=8.14.0=pypi_0
mpi=1.0=openmpi
mpmath=1.3.0=pypi_0
msgpack=1.0.4=pypi_0
munkres=1.1.4=py_0
mypy-extensions=0.4.3=pypi_0
nbformat=5.5.0=pypi_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
ntlm-auth=1.5.0=pypi_0
numba=0.55.2=pypi_0
numexpr=2.8.1=py37h2e5f0a9_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37h3b1a694_2
ogb=1.3.3=pyhd8ed1ab_0
omegaconf=2.2.3=pypi_0
open3d=0.16.1=pypi_0
opencv-python=4.7.0.68=pypi_0
openmesh=1.2.1=pypi_0
openmpi=4.0.2=hfa1e0ec_1
openssl=1.1.1t=hfd90126_0
outdated=0.2.2=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h743cdd8_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.10.1=pypi_0
pathtools=0.1.2=pypi_0
pbr=6.0.0=pypi_0
penaltymodel=1.0.2=pypi_0
performer-pytorch=1.1.4=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.1.1=pypi_0
pip=21.2.2=py37hecd8cb5_0
pkginfo=1.8.3=pypi_0
platformdirs=2.5.2=pypi_0
plotly=5.11.0=pypi_0
plucky=0.4.3=pypi_0
ply=3.11=pypi_0
potpourri3d=0.0.8=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.36=pyha770c72_0
protobuf=3.19.6=pypi_0
psutil=5.9.3=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pybind11=2.9.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=pypi_0
pyg=2.0.4=py37_torch_1.10.0_cpu
pygments=2.13.0=pypi_0
pykeops=2.1=pypi_0
pylatexenc=2.10=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyqubo=1.4.0=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.13=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.10.2=py3.7_0
pytorch-cluster=1.6.0=py37_torch_1.10.0_cpu
pytorch-scatter=2.0.9=py37_torch_1.10.0_cpu
pytorch-sparse=0.6.13=py37_torch_1.10.0_cpu
pytorch-spline-conv=1.2.1=py37_torch_1.10.0_cpu
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.3.0=pypi_0
pywin32-ctypes=0.2.0=pypi_0
pyyaml=6.0=py37hca72f7f_1
pyzmq=19.0.2=py37hf1e22d8_2
qiskit=0.43.3=pypi_0
qiskit-aer=0.12.2=pypi_0
qiskit-ibmq-provider=0.20.2=pypi_0
qiskit-terra=0.24.2=pypi_0
quantus=0.3.2=pypi_0
qubovert=1.2.5=pypi_0
ray=1.13.0=pypi_0
rdkit-pypi=2022.9.1=pypi_0
readline=8.1.2=hca72f7f_1
readme-renderer=37.2=pypi_0
rectangle-packer=2.0.1=pypi_0
requests=2.28.1=pypi_0
requests-ntlm=1.1.0=pypi_0
requests-toolbelt=0.10.0=pypi_0
rfc3986=2.0.0=pypi_0
rich=12.6.0=pypi_0
robust-laplacian=0.2.4=pypi_0
rustworkx=0.13.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=py37hae1ba45_1
scipy=1.7.3=py37h8c7af03_0
seaborn=0.12.2=pypi_0
sentry-sdk=1.10.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=61.2.0=py37hecd8cb5_0
shared-memory38=0.1.3=pypi_0
shortuuid=1.0.9=pypi_0
singledispatchmethod=1.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sparse=0.13.0=pypi_0
sqlite=3.38.3=h707629a_0
stevedore=3.5.2=pypi_0
symengine=0.9.2=pypi_0
sympy=1.10.1=pypi_0
tabulate=0.8.9=pypi_0
tenacity=8.1.0=pypi_0
tensorboardx=2.5.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=pypi_0
torch-summary=1.4.5=pypi_0
torchaudio=0.11.0=pypi_0
torchdiffeq=0.2.2=pyhd8ed1ab_0
torchmetrics=0.10.2=pypi_0
torchvision=0.11.3=pypi_0
tornado=6.1=py37h9ed2024_0
tqdm=4.64.0=py37hecd8cb5_0
traitlets=5.8.0=pypi_0
twine=4.0.1=pypi_0
typed-ast=1.5.5=pypi_0
typing-extensions=4.2.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.12=pypi_0
virtualenv=20.14.1=pypi_0
wandb=0.13.4=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
websockets=11.0.3=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.14.1=pypi_0
xz=5.2.5=hca72f7f_1
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=he49afe7_1
zipp=3.9.0=pypi_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
